refactor(compiler): inline unsafe-jit policy into JIT compiler
The unsafe-jit Trusted Types policy is only used in the JIT compiler.
To minimize the risk of the policy being used in inappropriate contexts,
which could lead to XSS vulnerabilities, move the policy directly into
the JIT compiler module.
